Better formatting for object references in console message
