+++ src/hub.js
@@ -1,10 +1,10 @@
+import "./utils/record-log"; // included to capture earliest log entries
 import {
   getCurrentHubId,
   updateVRHudPresenceCount,
   updateSceneCopresentState,
   createHubChannelParams
 } from "./utils/hub-utils";
-import "./utils/debug-log";
 import configs from "./utils/configs";
 import "./utils/theme";
 import "@babel/polyfill";
